# spec/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.GetResponse:
    properties:
      config:
        $ref: '#/definitions/models.ModelConfig'
      id:
        type: string
      model_id:
        type: string
      name:
        type: string
      provider_id:
        type: string
      type:
        $ref: '#/definitions/models.ModelType'
    type: object
...
  providers.CreateRequest:
    properties:
      client_type:
        type: string
      config:
        additionalProperties: {}
        type: object
      icon:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
    required:
    - client_type
    - name
    type: object
  providers.GetResponse:
    properties:
      client_type:
        type: string
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      enable:
        type: boolean
      icon:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      updated_at:
        type: string
    type: object
  providers.ImportModelsResponse:
    properties:
      created:
        type: integer
      models:
        items:
          type: string
        type: array
      skipped:
        type: integer
    type: object
...
  providers.UpdateRequest:
    properties:
      client_type:
        type: string
      config:
        additionalProperties: {}
        type: object
      enable:
        type: boolean
      icon:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /providers:
    get:
      consumes:
      - application/json
      description: Get a list of all configured LLM providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/providers.GetResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all LLM providers
      tags:
      - providers
    post:
      consumes:
      - application/json
      description: Create a new LLM provider configuration
      parameters:
      - description: Provider configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/providers.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/providers.GetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new LLM provider
      tags:
      - providers
  /providers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a provider configuration
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete provider
      tags:
      - providers
    get:
      consumes:
      - application/json
      description: Get a provider configuration by its ID
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/providers.GetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get provider by ID
      tags:
      - providers
    put:
      consumes:
      - application/json
      description: Update an existing provider configuration
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated provider configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/providers.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/providers.GetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update provider
      tags:
      - providers
  /providers/{id}/import-models:
    post:
      consumes:
      - application/json
      description: Fetch models from provider's /v1/models endpoint and import them
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/providers.ImportModelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Import models from provider
      tags:
      - providers
  /providers/{id}/models:
    get:
      description: Get models for a provider by id, optionally filtered by type
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Model type (chat, embedding)
        in: query
        name: type
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List provider models
      tags:
      - providers
...
  /providers/{id}/test:
    post:
      consumes:
      - application/json
      description: Probe a provider's base URL to check reachability, supported client
        types, and embedding support
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/providers.TestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Test provider connectivity
      tags:
      - providers
...
  /speech-models:
    get:
      description: List all models of type 'speech' (filtered view of unified models
        table)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/audio.SpeechModelResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all speech models
      tags:
      - speech-models
  /speech-models/{id}:
    get:
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audio.SpeechModelResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a speech model
      tags:
      - speech-models
    put:
      consumes:
      - application/json
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      - description: Model update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/audio.UpdateSpeechModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audio.SpeechModelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update a speech model
      tags:
      - speech-models
  /speech-models/{id}/capabilities:
    get:
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audio.ModelCapabilities'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get speech model capabilities
      tags:
      - speech-models
  /speech-models/{id}/test:
    post:
      consumes:
      - application/json
      description: Synthesize text using a specific model's config and return audio
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      - description: Text to synthesize
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/audio.TestSynthesizeRequest'
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Audio data
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Test speech model synthesis
      tags:
      - speech-models
  /speech-providers:
    get:
      description: List providers that support speech (filtered view of unified providers
        table)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/audio.SpeechProviderResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List speech providers
      tags:
      - speech-providers
  /speech-providers/{id}:
    get:
      description: Get a speech provider with masked config values
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audio.SpeechProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get speech provider
      tags:
      - speech-providers
  /speech-providers/{id}/import-models:
    post:
      consumes:
      - application/json
      description: Fetch models using the configured speech provider and import them
        into the unified models table
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audio.ImportModelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Import speech models from provider
      tags:
      - speech-providers
  /speech-providers/{id}/models:
    get:
      description: List models of type 'speech' for a specific speech provider
      parameters:
      - description: Provider ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/audio.SpeechModelResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List speech models by provider
      tags:
      - speech-providers
...
  /transcription-models:
    get:
      description: List all models of type 'transcription' (filtered view of unified
        models table)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/audio.TranscriptionModelResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all transcription models
      tags:
      - transcription-models
...
  /transcription-models/{id}/capabilities:
    get:
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audio.ModelCapabilities'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get transcription model capabilities
      tags:
      - transcription-models
...
swagger: "2.0"